# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Octo Protocol API
  version: 1.0.0
  description: >
    OpenAPI contract for the /v1 Octo Protocol routes.


    Every `amount_stroops` field below is a JSON `integer` and stays that way in this version —
    that is unchanged by issue #215. Storage now also carries an arbitrary-precision
    `amount_base_units NUMERIC(78,0)` column internally (needed so a future, non-Stellar chain
    whose amounts don't fit 64 bits can be represented exactly), but nothing about that is exposed
    on these schemas yet. The public cutover to a string-typed amount field is a separate,
    versioned breaking change tracked in issue #227 — do not assume `amount_stroops` stays numeric
    across a future major version.
servers:
  - url: /
paths:
...
  /v1/wallets/{id}/payment-links:
    post:
      summary: Create a payment link
      operationId: createPaymentLink
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
                  nullable: true
                image_url:
                  type: string
                  nullable: true
                redirect_url:
                  type: string
                  nullable: true
                  description: >
                    Where to send the payer's browser after their payment is confirmed. Octo
                    appends `?status=success&payment_id=<id>&slug=<slug>` when redirecting.
                amount_usdc_stroops:
                  type: integer
                  nullable: true
                  description: Fixed amount in USDC stroops (7dp). Omit/null for a flexible-amount link.
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentLinkResponse'
    get:
      summary: List payment links
      operationId: listPaymentLinks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
        - name: before
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListPaymentLinksResponse'
  /v1/wallets/{id}/payment-links/{link_id}:
    get:
      summary: Get a payment link
      operationId: getPaymentLink
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: link_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentLinkResponse'
    put:
      summary: Activate or deactivate a payment link
      operationId: setPaymentLinkActive
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: link_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - active
              properties:
                active:
                  type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentLinkResponse'
  /v1/wallets/{id}/payment-links/{link_id}/payments:
    get:
      summary: List payments recorded against a payment link
      description: Owner-authenticated only — includes payer name/email.
      operationId: listPaymentLinkPayments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: link_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
        - name: before
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListPaymentLinkPaymentsResponse'
  /v1/pay/{slug}:
    get:
      summary: Get a public payment link
      description: Public, no auth. Used by the hosted checkout page.
      operationId: getPublicPaymentLink
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicPaymentLinkResponse'
  /v1/pay/{slug}/intent:
    post:
      summary: Create a payment intent against a public payment link
      description: Public, no auth, rate-limited by IP.
      operationId: createPaymentIntent
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                payer_name:
                  type: string
                  nullable: true
                payer_email:
                  type: string
                  nullable: true
                amount_usdc_stroops:
                  type: integer
                  nullable: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntentResponse'
  /v1/pay/{slug}/payments/{payment_id}:
    get:
      summary: Poll a payment's status
      description: Public, no auth.
      operationId: getPaymentStatus
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: payment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentStatusResponse'
  /v1/pay/{slug}/signing-info:
    get:
      summary: Get Stellar signing info for the payer's own account
      description: Public, no auth.
      operationId: publicSigningInfo
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: account
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSigningInfoResponse'
  /v1/pay/{slug}/submit-signed:
    post:
      summary: Relay a payer-signed transaction to Horizon
      description: Public, no auth.
      operationId: submitPayment
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                transaction_xdr:
                  type: string
                payment_id:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitPaymentResponse'
...
components:
  schemas:
...
    PaymentLinkResponse:
      type: object
      required:
        - statusCode
        - message
        - data
      properties:
        statusCode:
          type: integer
        message:
          type: string
        data:
          type: object
          required:
            - id
            - slug
            - name
            - active
            - collected_usdc_stroops
            - created_at
            - url
          properties:
            id:
              type: string
              format: uuid
            slug:
              type: string
            name:
              type: string
            description:
              type: string
              nullable: true
            image_url:
              type: string
              nullable: true
            redirect_url:
              type: string
              nullable: true
            amount_usdc_stroops:
              type: integer
              nullable: true
              description: Null means flexible — the payer chooses the amount.
            active:
              type: boolean
            collected_usdc_stroops:
              type: integer
            created_at:
              type: string
              format: date-time
            url:
              type: string
              description: Full hosted checkout URL, e.g. `https://app.octo.dev/pay/ab12cd34ef`.
    ListPaymentLinksResponse:
      type: object
      required:
        - statusCode
        - message
        - data
      properties:
        statusCode:
          type: integer
        message:
          type: string
        data:
          type: object
          required:
            - data
          properties:
            data:
              type: array
              items:
                type: object
                required:
                  - id
                  - slug
                  - name
                  - active
                  - collected_usdc_stroops
                  - created_at
                  - url
                properties:
                  id:
                    type: string
                    format: uuid
                  slug:
                    type: string
                  name:
                    type: string
                  description:
                    type: string
                    nullable: true
                  image_url:
                    type: string
                    nullable: true
                  redirect_url:
                    type: string
                    nullable: true
                  amount_usdc_stroops:
                    type: integer
                    nullable: true
                  active:
                    type: boolean
                  collected_usdc_stroops:
                    type: integer
                  created_at:
                    type: string
                    format: date-time
                  url:
                    type: string
            next_cursor:
              type: string
              format: uuid
              nullable: true
...
    PaymentIntentResponse:
      type: object
      required:
        - statusCode
        - message
        - data
      properties:
        statusCode:
          type: integer
        message:
          type: string
        data:
          type: object
          required:
            - payment_id
            - deposit_address
            - amount_usdc_stroops
          properties:
            payment_id:
              type: string
              format: uuid
            deposit_address:
              type: string
            amount_usdc_stroops:
              type: integer
...